requirements.txt
oci
oracledb
certifi
boto3
django-storages
awscli

asgiref==3.7.2
Django~=5.0.4
sqlparse==0.5.0
Pillow==10.3.0
python-decouple==3.8